requirements.txt
Fastapi
google-genai
uvicorn
pydub
python-dotenv
